refactor(mobile): more reorganization/cleanup
