+++ apps/mobile/src/features/create/components/latest-gallery-image.tsx
